requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
DateTime==5.5
filelock==3.16.1
Flask==3.0.3
future==1.0.0
gevent==24.10.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
PySocks==1.7.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
setuptools==75.1.0
tabulate==0.9.0
tldextract==5.1.2
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.1.0
